Simple_Tasks_Web_App/requirements.txt
pandas==2.2.3
streamlit==1.42.2   
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10